requirements.txt
apache-airflow
tweepy
pandas
python-dotenv
psycopg2-binary
SQLAlchemy